requirements.txt
openai
python-multipart
fastapi
pydantic
uvicorn
python-dotenv
pinecone-client